Refactor class layout
